+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add support for Windows7 by fixing rust to 1.77.2
 - Add build info to app, available via command line with --version or in About menu in the UI
 - Add rust, typescript, javascript, markdown code formating and linting
+- Add translation from localhost to 127.0.0.1 to speed up connections due to Windows DNS
 
 ### Removed